Added a module for querying Proxmox to gather info on VMs.
The main features of this module allow the user to query a Proxmox cluster on the status of VMs.

The user can query the cluster for VMs using the following arguments:

node: "<node_name>" returns all VMs on the node.
vmid: "<vmid>" returns the specific VM.
name: "<vm_name>" return all VMs with that name.

If the user excludes the arguments node, vmid, and name then all VMs in the cluster will be returned.

The user can also decide what information is returned about the vm using the 'load_section' argument which accepts a list of sections.

load_sections:
  - all
  - config
  - snapshots
  - firewall
  - agent_network_info
  - agent_os_info